Add Makefile for easier contributing
This Makefile and the CONTRIBUTING.md are modified from the versions
included with Dist::Zilla::PluginBundle::Author::GSG.  They allow folks
who just want to add a small change to more easily contribute by making
it easier to run tests and to install dependencies.

To keep the new CONTRIBUTING docs around we need to stop overwiting it.